backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
PyMySQL
python-dotenv
Flask-JWT-Extended
Werkzeug
Flask-Cors